+++ src/main/java/fr/estecka/packrulemenus/DatapackHandler.java
@@ -1,22 +1,22 @@
 package fr.estecka.packrulemenus;
 
 import java.util.Collection;
-import net.minecraft.client.MinecraftClient;
-import net.minecraft.client.gui.screen.Screen;
-import net.minecraft.client.gui.screen.TitleScreen;
-import net.minecraft.client.gui.screen.pack.PackScreen;
-import net.minecraft.client.gui.widget.ButtonWidget;
-import net.minecraft.client.world.ClientWorld;
-import net.minecraft.resource.DataConfiguration;
-import net.minecraft.resource.ResourcePackManager;
-import net.minecraft.resource.featuretoggle.FeatureFlags;
-import net.minecraft.resource.featuretoggle.FeatureSet;
-import net.minecraft.server.integrated.IntegratedServer;
-import net.minecraft.text.MutableText;
-import net.minecraft.text.Text;
-import net.minecraft.util.Formatting;
-import net.minecraft.util.Identifier;
-import net.minecraft.util.WorldSavePath;
+import net.minecraft.ChatFormatting;
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.gui.components.Button;
+import net.minecraft.client.gui.screens.Screen;
+import net.minecraft.client.gui.screens.TitleScreen;
+import net.minecraft.client.gui.screens.packs.PackSelectionScreen;
+import net.minecraft.client.multiplayer.ClientLevel;
+import net.minecraft.client.server.IntegratedServer;
+import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.MutableComponent;
+import net.minecraft.resources.Identifier;
+import net.minecraft.server.packs.repository.PackRepository;
+import net.minecraft.world.flag.FeatureFlagSet;
+import net.minecraft.world.flag.FeatureFlags;
+import net.minecraft.world.level.WorldDataConfiguration;
+import net.minecraft.world.level.storage.LevelResource;
 import fr.estecka.packrulemenus.gui.GenericWarningScreen;
 import fr.estecka.packrulemenus.mixin.IMinecraftServerMixin;
 
@@ -26,45 +26,45 @@ public class DatapackHandler
 
 	private final Screen parent;
 	private final IntegratedServer server;
-	private final ResourcePackManager manager;
+	private final PackRepository manager;
 	private final Collection<String> rollback;
-	static private final MinecraftClient client = MinecraftClient.getInstance();
+	static private final Minecraft client = Minecraft.getInstance();
 
 	private DatapackHandler(Screen parent, IntegratedServer server){
 		this.parent = parent;
 		this.server = server;
-		this.manager = server.getDataPackManager();
-		this.rollback = manager.getEnabledIds();
+		this.manager = server.getPackRepository();
+		this.rollback = manager.getSelectedIds();
 	}
 
-	static public ButtonWidget CreateButton(Screen parent){
+	static public Button CreateButton(Screen parent){
 		return PackRuleMod.DisabledButton(BUTTON_TEXT)
-			.orElseGet(()->new DatapackHandler(parent, client.getServer()).CreateButton())
+			.orElseGet(()->new DatapackHandler(parent, client.getSingleplayerServer()).CreateButton())
 			;
 	}
 
-	public ButtonWidget CreateButton(){
-		return ButtonWidget.builder(
-			Text.translatable(BUTTON_TEXT),
+	public Button CreateButton(){
+		return Button.builder(
+			Component.translatable(BUTTON_TEXT),
 			__->client.setScreen( this.CreateScreen() )
 		).build();
 	}
 
-	public PackScreen CreateScreen(){
-		return new PackScreen(
-			server.getDataPackManager(),
+	public PackSelectionScreen CreateScreen(){
+		return new PackSelectionScreen(
+			server.getPackRepository(),
 			manager -> { HandleDatapackRefresh(manager, rollback); },
-			server.getSavePath(WorldSavePath.DATAPACKS),
-			Text.translatable("dataPack.title")
+			server.getWorldPath(LevelResource.DATAPACK_DIR),
+			Component.translatable("dataPack.title")
 		);
 	}
 
-	private void	HandleDatapackRefresh(final ResourcePackManager manager, Collection<String> rollback){
-		FeatureSet neoFeatures = manager.getRequestedFeatures();
-		FeatureSet oldFeatures = server.getSaveProperties().getEnabledFeatures();
+	private void	HandleDatapackRefresh(final PackRepository manager, Collection<String> rollback){
+		FeatureFlagSet neoFeatures = manager.getRequestedFeatureFlags();
+		FeatureFlagSet oldFeatures = server.getWorldData().enabledFeatures();
 
 		if (!neoFeatures.equals(oldFeatures)){
-			boolean isExperimental = FeatureFlags.isNotVanilla(neoFeatures);
+			boolean isExperimental = FeatureFlags.isExperimental(neoFeatures);
 			boolean wasVanillaRemoved = oldFeatures.contains(FeatureFlags.VANILLA) && !neoFeatures.contains(FeatureFlags.VANILLA);
 			ShowFeatureWarning(isExperimental, wasVanillaRemoved);
 		}
@@ -80,38 +80,38 @@ public class DatapackHandler
 /******************************************************************************/
 
 	private void	ApplyFlags(){
-		FeatureSet features = manager.getRequestedFeatures();
+		FeatureFlagSet features = manager.getRequestedFeatureFlags();
 
 		String featureNames = "";
-		for (Identifier id : FeatureFlags.FEATURE_MANAGER.toId(features))
+		for (Identifier id : FeatureFlags.REGISTRY.toNames(features))
 			featureNames += id.toString()+", ";
 		PackRuleMod.LOGGER.info("Reloading packs with features: {}", featureNames);
 
-		server.getSaveProperties().updateLevelInfo(new DataConfiguration(IMinecraftServerMixin.callCreateDataPackSettings(manager, true), features));		
+		server.getWorldData().setDataConfiguration(new WorldDataConfiguration(IMinecraftServerMixin.callGetSelectedPacks(manager, true), features));		
 	}
 
 	private void SaveAndQuit(){
 		this.ApplyFlags();
-		this.server.stop(false);
-		if (client.world != null)
-			client.world.disconnect(ClientWorld.QUITTING_MULTIPLAYER_TEXT);
+		this.server.halt(false);
+		if (client.level != null)
+			client.level.disconnect(ClientLevel.DEFAULT_QUIT_MESSAGE);
 		client.disconnectWithSavingScreen();
 		client.setScreen(new TitleScreen());
 	}
 
 	private void	ReloadPacks(){
-		client.inGameHud.getChatHud().addMessage(Text.translatable("commands.reload.success"));
+		client.gui.getChat().addServerSystemMessage(Component.translatable("commands.reload.success"));
 
-		server.reloadResources(manager.getEnabledIds()).exceptionally(e -> {
+		server.reloadResources(manager.getSelectedIds()).exceptionally(e -> {
 			PackRuleMod.LOGGER.error("{}", e);
-			client.inGameHud.getChatHud().addMessage(Text.translatable("commands.reload.failure").formatted(Formatting.RED));
+			client.gui.getChat().addServerSystemMessage(Component.translatable("commands.reload.failure").withStyle(ChatFormatting.RED));
 			return null;
 		});
 		client.setScreen(parent);
 	}
 
 	private void Rollback(){
-		this.manager.setEnabledProfiles(rollback);
+		this.manager.setSelected(rollback);
 		client.setScreen(parent);
 	}
 
@@ -122,9 +122,9 @@ public class DatapackHandler
 
 	public void	ShowConfirmationScreen(){
 		client.setScreen(new GenericWarningScreen(
-			Text.translatable("packrulemenus.warning.packConfirmation.title"),
-			Text.translatable("packrulemenus.warning.packConfirmation.message"),
-			Text.translatable("packrulemenus.warning.packConfirmation.checkbox"),
+			Component.translatable("packrulemenus.warning.packConfirmation.title"),
+			Component.translatable("packrulemenus.warning.packConfirmation.message"),
+			Component.translatable("packrulemenus.warning.packConfirmation.checkbox"),
 			false,
 			checked -> { if(checked) SaveAndQuit(); else ReloadPacks(); },
 			this::Rollback
@@ -136,14 +136,14 @@ public class DatapackHandler
 	 * experimental features. This will be false upon removing all features.
 	 */
 	public void	ShowFeatureWarning(boolean isExperimental, boolean wasVanillaRemoved){
-		MutableText msg = Text.translatable("packrulemenus.warning.featureflag.message");
+		MutableComponent msg = Component.translatable("packrulemenus.warning.featureflag.message");
 		if (isExperimental)
-			msg.append("\n\n").append(Text.translatable("selectWorld.experimental.message"));
+			msg.append("\n\n").append(Component.translatable("selectWorld.experimental.message"));
 
 		client.setScreen(new GenericWarningScreen(
-			Text.translatable("packrulemenus.warning.featureflag.title"),
+			Component.translatable("packrulemenus.warning.featureflag.title"),
 			msg,
-			Text.translatable("packrulemenus.warning.featureflag.checkbox"),
+			Component.translatable("packrulemenus.warning.featureflag.checkbox"),
 			true,
 			checked -> { if(wasVanillaRemoved) ShowVanillaWarning(); else if (checked) SaveAndQuit(); },
 			this::Rollback
@@ -152,9 +152,9 @@ public class DatapackHandler
 
 	public void	ShowVanillaWarning(){
 		client.setScreen(new GenericWarningScreen(
-			Text.translatable("packrulemenus.warning.vanillapack.title"),
-			Text.translatable("packrulemenus.warning.vanillapack.message"),
-			Text.translatable("packrulemenus.warning.vanillapack.checkbox"),
+			Component.translatable("packrulemenus.warning.vanillapack.title"),
+			Component.translatable("packrulemenus.warning.vanillapack.message"),
+			Component.translatable("packrulemenus.warning.vanillapack.checkbox"),
 			true,
 			checked -> { if(checked) SaveAndQuit(); },
 			this::Rollback